+++ bevy-butler/tests/add_system/generic_pipe.rs
@@ -2,7 +2,6 @@ use std::fmt::Display;
 
 use bevy::prelude::*;
 use bevy_butler::*;
-use bevy_log::prelude::*;
 
 use crate::common::log_plugin;
 
@@ -10,7 +9,7 @@ use crate::common::log_plugin;
 struct MyPlugin;
 
 #[derive(Resource)]
-#[add_resource(plugin = MyPlugin, init = GenericRes(5u8))]
+#[insert_resource(plugin = MyPlugin, init = GenericRes(5u8))]
 struct GenericRes<T>(T);
 
 fn generic_pipe<T: 'static + Sync + Send + Display>(res: Res<GenericRes<T>>) -> String {