+++ bevy-butler/tests/add_system/system_expr_schedule.rs
@@ -6,7 +6,6 @@ use bevy::{prelude::*, time::TimePlugin};
 use bevy_app::ScheduleRunnerPlugin;
 use bevy_butler::*;
 use bevy_state::{app::StatesPlugin, prelude::*};
-use bevy_log::prelude::*;
 
 use crate::common::log_plugin;