+++ bevy-butler/tests/insert_resource/resource.rs
@@ -7,24 +7,24 @@ use wasm_bindgen_test::wasm_bindgen_test;
 use crate::common::log_plugin;
 
 #[derive(Resource, Default)]
-#[add_resource(plugin = MyPlugin, generics = <bool>)]
-struct Marker<T>(T);
+#[insert_resource(plugin = MyPlugin)]
+struct Marker(bool);
 
 #[derive(Resource)]
-#[add_resource(plugin = MyPlugin, init = Message("Hello, world!".to_string()))]
-struct Message<T>(T);
+#[insert_resource(plugin = MyPlugin, init = Message("Hello, world!".to_string()))]
+struct Message(String);
 
 #[butler_plugin]
 struct MyPlugin;
 
 #[add_system(plugin = MyPlugin, schedule = Startup)]
-fn get_and_print_message(message: Res<Message<String>>, mut marker: ResMut<Marker<bool>>) {
+fn get_and_print_message(message: Res<Message>, mut marker: ResMut<Marker>) {
     info!("Resource message: {}", message.0);
     marker.0 = true;
 }
 
 #[add_system(plugin = MyPlugin, schedule = PostStartup)]
-fn assert_marker(marker: Res<Marker<bool>>) {
+fn assert_marker(marker: Res<Marker>) {
     assert!(marker.0);
 }