+++ bevy-butler/tests/insert_resource/resource_enum.rs
@@ -1,7 +1,6 @@
 use bevy::prelude::*;
 use bevy_butler::*;
 use wasm_bindgen_test::wasm_bindgen_test;
-use bevy_log::prelude::*;
 
 use crate::common::log_plugin;
 
@@ -10,7 +9,7 @@ struct MyPlugin;
 
 #[allow(dead_code)]
 #[derive(Resource, Default, Debug)]
-#[add_resource(plugin = MyPlugin)]
+#[insert_resource(plugin = MyPlugin)]
 enum Message {
     Variant1,
     #[default]